backend/ai/OCR_AI/requirements.txt
filelock==3.20.0
fsspec==2025.12.0
Jinja2==3.1.6
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.6.1
numpy==2.3.5
pillow==12.0.0
setuptools==70.2.0
sympy==1.14.0
torch
torchaudio
torchvision
Flask
opencv-python
typing_extensions==4.15.0
